requirements.txt
flask
opencv-python
pytesseract
matplotlib
sqlalchemy
psycopg2-binary
PyJWT
sqlalchemy-serializer
